requirements.txt
Epiphyte
pycrypto >= 2.6
requests >= 2.0